Address task list review feedback
